% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BF_methods.R, R/BFvariances.R
\name{BF}
\alias{BF}
\alias{BF.vtest}
\title{BF test for a univariate normal linear regression type model}
\usage{
BF(x, hypothesis, prior, parameter, ...)

\method{BF}{vtest}(x, hypothesis = NULL, prior = NULL,
  parameter = NULL, ...)
}
\arguments{
\item{x}{An R object containing the outcome of a statistical analysis.}

\item{hypothesis}{A character string containing the informative hypotheses to
evaluate. Defaults to NULL, which will
result in an exploratory analysis.}

\item{prior}{Prior, Default: 'default'}

\item{parameter}{Joris, wat is dit?}

\item{...}{Parameters passed to and from other functions.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
