% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_estimator.R
\name{BF_Gaussian}
\alias{BF_Gaussian}
\title{Bayes factor testing based on Gaussian approximations}
\usage{
BF_Gaussian(estimate, sigma, n, hypothesis = NULL, prior = NULL)
}
\arguments{
\item{estimate}{A named vector containing the estimates of the parameters that
are tested.}

\item{sigma}{The error covariance matrix of the estimates.}

\item{n}{The sample size.}

\item{hypothesis}{A character string containing the informative hypotheses to
evaluate. The default is NULL, which will result in an exploratory analysis.}

\item{prior}{A vector specifying the prior probabilities of the hypotheses.
The default is NULL which will specify equal prior probabilities.}
}
\value{
The output is an object of class \code{BF}. The object has elements:
BFtu_exploratory, PHP_exploratory, BFtu_confirmatory, PHP_confirmatory,
BFmatrix_confirmatory, BFtable_confirmatory, BFtu_main, PHP_main,
BFtu_interaction, PHP_interaction, prior, hypotheses, estimates, model, call.
}
\description{
The function \code{BF_Gaussian} can be used for hypothesis testing
and model selection using the Bayes factor using a Gaussian approximation of the
likelihood function. By default exploratory hypothesis tests are
performed of whether each model parameter equals zero, is negative, or is positive.
Confirmatory hypothesis tests can be executed by specifying hypotheses with
equality and/or order constraints on the parameters of interest.
}
\details{
The function requires a named vector \code{estimate} together with
its error covariance matrix \code{sigma} and the sample size \code{n} of the
data that were used to get the estimates of the fitted model.
}
\examples{

# load R package BFpack
library(BFpack)

# Logistic regression
fit1 <- glm(sent ~ ztrust + zfWHR + zAfro + glasses + attract + maturity +
   tattoos, family = binomial(), data = wilson)
estimate <- coef(fit1)
sigma <- vcov(fit1)
n <- nobs(fit1)

BF1 <- BF_Gaussian(estimate, sigma, n, hypothesis="ztrust > 0 & zfWHR=zAfro= 0")
summary(BF1)

#the result is the same as when running
BF(fit1, hypothesis = "ztrust > 0 & zfWHR=zAfro= 0")
}
\references{
Mulder, J., Gu, X., A. Tomarken, F. Boing-Messing,
J.A.O.C. Olsson-Collentine, Marlyne Bosman-Meyerink, D.R. Williams, J. Menke,
J.-P. Fox, Y. Rosseel, E.J. Wagenmakers, H. Hoijtink., van Lissa, C. (submitted). BFpack:
Flexible Bayes Factor Testing of Scientific Theories in R.
}
