% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_orderconstraints.R
\name{bic_oc}
\alias{bic_oc}
\title{Compute the BIC of an order-constrained model.}
\usage{
bic_oc(object, constraints = NULL, complement = F, numdraws = 1e+05)
}
\arguments{
\item{object}{A fitted model object, such as a glm-object}

\item{constraints}{A string specifying order constraints on certain effects of the modeling object. If the value is NULL the ordinary BIC is computed.}

\item{complement}{A logical scalar that specifies if the order-constrained subspace is considered (FALSE) or its complement (TRUE). Default is FALSE.}

\item{numdraws}{A number specifying the number of draws that are used to compute the posterior and prior probability that the order constraints hold. This setting is only when the number of \code{constraints} is larger than the rank of the coefficient matrix corresponding to the \code{constraints}.}
}
\value{
Return a list of which the order-constrained BIC of modeling \code{object} with \code{constraints} based on the local unit-information prior as first element, the posterior probability that the constraints hold as second element, and the prior probability that the constraints hold as third element. If \code{complement} is TRUE then the complement of the order-constrained subspace is considered.
}
\description{
\code{bic_oc} computes the order-constrained BIC for a fitted model object (e.g., lm-object, a glm-object, or a coxph-object) with order constraints on certain effects.
}
\examples{
n <- 100
x1 <- rnorm(n)
x2 <- rnorm(n)
y <- 1 + .5 * x1 + 1 * x2 + rnorm(n)
df <- data.frame(y=y,x1=x1,x2=x2)
glm1 <- lm(y ~ 1 + x1 + x2, data=df)
# Compute the BIC of the fitted model `glm1' with order constraints that state that
# the effect of 'x2' on 'y' is larger than the effect of 'x1' on 'y', and both effects
# are assumed positive.
bic_oc(glm1,"x2 > x1 > 0")
}
