% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BF_cortest.R
\name{cor_test}
\alias{cor_test}
\title{Bayesian correlation analysis}
\usage{
cor_test(..., formula = NULL, iter = 5000)
}
\arguments{
\item{...}{matrices (or data frames) of dimensions \emph{n} (observations) by  \emph{p} (variables)
for different groups (in case of multiple matrices or data frames).}

\item{formula}{an object of class \code{\link[stats]{formula}}. This allows for including
control variables in the model (e.g., \code{~ education}).}

\item{iter}{number of iterations from posterior (default is 5000).}
}
\value{
list of class \code{cor_test}:
\itemize{
\item \code{meanF} posterior means of Fisher transform correlations
\item \code{covmF} posterior covariance matrix of Fisher transformed correlations
\item \code{correstimates} posterior estimates of correlation coefficients
\item \code{corrdraws} list of posterior draws of correlation matrices per group
\item \code{corrnames} names of all correlations
}
}
\description{
Estimate the unconstrained posterior for the correlations using a joint uniform prior.
}
\examples{

# Bayesian correlation analysis of the 6 variables in 'memory' object
# sample posterior
fit <- cor_test(BFpack::memory[,1:6])

# Bayesian correlation of variables in memory object in BFpack while controlling
# for the Group variable
# sample posterior
fit <- cor_test(BFpack::memory,formula = ~ Group)

# Bayesian correlation of variables in memory object in BFpack separate for the
# two different groups
# sample posterior
HC <- subset(BFpack::memory, Group == "HC")[,-7]
SZ <- subset(BFpack::memory, Group == "SZ")[,-7]
fit <- cor_test(HC,SZ)

}
