% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BF.cortest.R
\name{cor_test_continuous}
\alias{cor_test_continuous}
\title{Bayesian correlation analysis}
\usage{
cor_test_continuous(..., formula = NULL, iter = 5000, burnin = 3000)
}
\arguments{
\item{...}{matrices (or data frames) of dimensions \emph{n} (observations) by  \emph{p} (variables)
for different groups (in case of multiple matrices or data frames).}

\item{formula}{an object of class \code{\link[stats]{formula}}. This allows for including
control variables in the model (e.g., \code{~ education}).}

\item{iter}{number of iterations from posterior (default is 5000).}

\item{burnin}{number of iterations for burnin (default is 3000).}
}
\value{
list of class \code{cor_test}:
\itemize{
\item \code{meanF} posterior means of Fisher transform correlations
\item \code{covmF} posterior covariance matrix of Fisher transformed correlations
\item \code{correstimates} posterior estimates of correlation coefficients
\item \code{corrdraws} list of posterior draws of correlation matrices per group
\item \code{corrnames} names of all correlations
}
}
\description{
Estimate the unconstrained posterior for the correlations using a joint uniform prior.
}
\examples{
\donttest{

# Bayesian correlation analysis of the 6 variables in 'memory' object
# we consider a correlation analysis of the first three variable of the memory data.
#fit <- cor_test_continuous(BFpack::memory[,1:3])
}
}
