% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fcor.R
\docType{data}
\name{Fcor}
\alias{Fcor}
\title{Student t approximations of Fisher transformed correlations}
\format{
A data.frame with 3 columns.
}
\usage{
data(Fcor)
}
\description{
Approximated degrees of freedom and approximated scale of the Fisher transformed
correlations depending on the dimension of the vector of dependent variables P
based on a joint uniform prior.
}
\details{
\tabular{lll}{
   \strong{nu} \tab \code{numeric} \tab Approximated degrees of freedom\cr
   \strong{sigma} \tab \code{numeric} \tab Approximated scale\cr
   \strong{P} \tab \code{integer} \tab Dimension of vector of dependent variables\cr
}
}
\keyword{datasets}
