% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri.R
\docType{data}
\name{fmri}
\alias{fmri}
\title{fmri Data}
\format{A data.frame with 14 rows and 5 columns.}
\usage{
data(fmri)
}
\description{
Decription of origin of the data.
}
\details{
\tabular{lll}{
   \strong{Face} \tab \code{numeric} \tab Participant face measurement\cr
   \strong{Vehicle} \tab \code{numeric} \tab Participant vehicle measurement\cr
   \strong{Superfical} \tab \code{numeric} \tab Superficial layes\cr
   \strong{Middle} \tab \code{numeric} \tab Middle layer\cr
   \strong{Bform} \tab \code{numeric} \tab Deep layer\cr
}
}
\references{
REFERENCE FOR THE DATA.
}
\keyword{datasets}
