% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_location.R
\docType{data}
\name{same_location}
\alias{same_location}
\title{Same location event statistic}
\format{
dataframe (25 rows, 4 columns)

\tabular{lll}{
   \strong{same_location} \tab \code{integer} \tab Event statistic. Matrix with senders in the
   rows and receivers in the columns. The event statistic is 1 if sender and receiver have
   the same location and 0 otherwise. \cr
}
}
\usage{
data(same_location)
}
\description{
A matrix coding whether senders of events (in the rows) and receivers of events
(in the column) have the same location. Related to the 'events' data object,
that contains a relational event sequence, and the 'actors' object, that contains
information on the 25 actors involved in the relational event sequence.
}
\keyword{datasets}
