% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvprices.R
\docType{data}
\name{tvprices}
\alias{tvprices}
\title{Precision of the Anchor Influences the Amount of Adjustment}
\format{A data.frame with 59 rows and 3 columns.}
\source{
\href{https://doi.org/10.1111/j.1467-9280.2008.02057.x}{
doi:10.1111/j.1467-9280.2008.02057.x}
}
\usage{
data(tvprices)
}
\description{
Data from an experimental study where participants have to guess the price
of a plasma tv. There were two experimental conditions.
These data were simulated using the sufficient statistics from Janiszewski &
Uy (2008).
}
\details{
\tabular{lll}{
   \strong{price} \tab \code{numeric} \tab Participant z-scores of price\cr
   \strong{anchor} \tab \code{factor} \tab Participant anchor\cr
   \strong{motivation} \tab \code{factor} \tab motivation to change\cr
}
}
\references{
Janiszewski, C., & Uy, D. (2008). Precision of the anchor influences the amount of adjustment. Psychological Science, 19(2), 121–127. https://doi.org/10.1111/j.1467-9280.2008.02057.x
\href{https://doi.org/10.1111/j.1467-9280.2008.02057.x}{
doi:10.1111/j.1467-9280.2008.02057.x}
}
\keyword{datasets}
