% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFvariances.R
\name{var_test}
\alias{var_test}
\alias{var_test.default}
\alias{var_test.formula}
\title{Bartlett Test of Homogeneity of Variances}
\usage{
var_test(x, ...)

\method{var_test}{default}(x, g, ...)

\method{var_test}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{a numeric vector of data values, or a list of numeric data vectors
representing the respective samples, or fitted linear model objects
(inheriting from class "\code{lm}").}

\item{...}{Parameters passed to and from other functions.}

\item{g}{a vector or factor object giving the group for the corresponding
elements of \code{x}. Ignored if \code{x} is a list.}

\item{formula}{a formula of the form \code{lhs ~ rhs} where lhs gives the
data values and rhs the corresponding groups.}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in the formula
\code{formula}. By default the variables are taken from
environment(\code{formula}).}

\item{subset}{an optional vector specifying a subset of observations to be
used.}

\item{na.action}{a function which indicates what should happen when the data
contain \code{NA}s. Defaults to getOption("\code{na.action}").}
}
\value{
A list with class \code{"vtest"} containing the following
components:
\item{statistic}{Bartlett's K-squared test statistic.}
\item{parameter}{the degrees of freedom of the approximate chi-squared
distribution of the test statistic.}
\item{p.value}{the p-value for the test.}
\item{method}{the character string "\code{Bartlett test of homogeneity of
variances}".}
\item{data.name}{a character string giving the name(s) of
the data.}
\item{v}{a vector containing the sample variances for each factor level.}
\item{n}{a vector containing the sample sizes for each factor level.}
}
\description{
Performs Bartlett's test of the null that the variances in each
of the groups (samples) are the same.
}
\details{
If \code{x} is a list, its elements are taken as the samples or
fitted linear models to be compared for homogeneity of variances. In this
case, the elements must either all be numeric data vectors or fitted linear
model objects, \code{g} is ignored, and one can simply use
\code{bartlett.test(x)} to perform the test. If the samples are not yet
contained in a list, use \code{var_test(list(x, ...))}.

Otherwise, x must be a numeric data vector, and g must be a vector or factor
object of the same length as x giving the group for the corresponding
elements of x.
}
\section{BF var_test}{

In order to allow users to enjoy the functionality of BFpack with the
familiar
stats-function \code{bartlett.test}, we have had to make minor changes to the
function
bartlett.test.default. All rights to, and credit for, the function
\code{bartlett.test.default}
belongs to the R Core Team, as indicated in the original license below.
We make no claims to copyright and incur no liability with regard to the
changes implemented in var_test.

This the original copyright notice by the R core team:
File src/library/stats/R/t_test.R
Part of the R package, https://www.R-project.org

Copyright (C) 1995-2015 The R Core Team

This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 2 of the License, or
(at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

A copy of the GNU General Public License is available at
https://www.R-project.org/Licenses/
}

\examples{
require(graphics)

plot(count ~ spray, data = InsectSprays)
var_test(InsectSprays$count, InsectSprays$spray)
var_test(count ~ spray, data = InsectSprays)

}
\references{
Bartlett, M. S. (1937). Properties of sufficiency and statistical
tests. Proceedings of the Royal Society of London Series A 160, 268–282.
doi: 10.1098/rspa.1937.0109.
}
\seealso{
\code{\link{var.test}} for the special case of comparing variances
in two samples from normal distributions; \code{\link{fligner.test}} for a
rank-based (nonparametric) k-sample test for homogeneity of variances;
\code{\link{ansari.test}} and \code{\link{mood.test}} for two rank based
two-sample tests for difference in scale.
}
