% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri.R
\docType{data}
\name{fmri}
\alias{fmri}
\title{fMRI data}
\format{
A data.frame with 13 rows and 6 columns.
}
\usage{
data(fmri)
}
\description{
fMRI data assessing relation between individual differences in the ability to recognize
faces and cars and thickness of the superficial, middle, and deep layers of the
fusiform face area, as assessed by high-resolution fMRI recognition (Williams et al, 2019, under review)
}
\details{
\tabular{lll}{
  \strong{Subject}\tab\code{numeric}\tab Particicpant ID number\cr
   \strong{Face} \tab \code{numeric} \tab Standardized score on face recognition battery\cr
   \strong{Vehicle} \tab \code{numeric} \tab Standardized score on vehicle recognition battery\cr
   \strong{Superficial} \tab \code{numeric} \tab Depth in mm of superficial layer of FFA\cr
   \strong{Middle} \tab \code{numeric} \tab Depth in mm of middle layer of FFA\cr
   \strong{Bform} \tab \code{numeric} \tab Depth in mm of deep layer of FFA\cr
}
}
\references{
McGuigin, R.W., Newton, A.T., Tamber-Rosenau, B., Tomarken, A.J, & Gauthier, I. (under review). Thickness of deep layers in the fusiform face area predicts face recognition.
}
\keyword{datasets}
