% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFvar.R
\name{var_test}
\alias{var_test}
\alias{var_test.default}
\title{Bartlett Test of Homogeneity of Variances}
\usage{
var_test(x, ...)

\method{var_test}{default}(x, g, ...)
}
\description{
Bartlett Test of Homogeneity of Variances
}
